requirements.txt
numpy
scipy
h5py
pyyaml
torch

